Recent Contributors section - refactor followup (#231)
* followup refactor of utils.

* fix indentation
